Day #1034: Bugfixes/experiments/notes.

+++ side_experiments/factor/solution/action_node.cpp
@@ -51,8 +51,8 @@ void ActionNode::clean() {
 }
 
 void ActionNode::measure_update() {
-	this->average_score = this->sum_score / this->num_measure;
-	this->average_instances_per_run = this->num_measure / MEASURE_ITERS;
+	this->average_score = this->sum_score / (double)this->num_measure;
+	this->average_instances_per_run = (double)this->num_measure / MEASURE_ITERS;
 }
 
 void ActionNode::save(ofstream& output_file) {
