+++ side_experiments/factor/solution/scope_node.cpp
@@ -72,8 +72,8 @@ void ScopeNode::clean() {
 }
 
 void ScopeNode::measure_update() {
-	this->average_score = this->sum_score / this->num_measure;
-	this->average_instances_per_run = this->num_measure / MEASURE_ITERS;
+	this->average_score = this->sum_score / (double)this->num_measure;
+	this->average_instances_per_run = (double)this->num_measure / MEASURE_ITERS;
 }
 
 void ScopeNode::save(ofstream& output_file) {
